Change the client structure
