+++ src/main/java/oop/course/client/GUIOrchestrator.java
@@ -3,6 +3,8 @@ package oop.course.client;
 import com.googlecode.lanterna.gui2.MultiWindowTextGUI;
 import com.googlecode.lanterna.gui2.WindowBasedTextGUI;
 import com.googlecode.lanterna.screen.Screen;
+import oop.course.client.actions.Action;
+import oop.course.client.views.*;
 
 import java.io.IOException;
