--- src/main/java/oop/course/client/Request.java
@@ -1,13 +1,0 @@
-package oop.course.client;
-
-import java.io.PrintWriter;
-
-public interface Request {
-
-    enum Method {
-        GET,
-        POST
-    }
-
-    void send(PrintWriter printWriter);
-}
