--- src/main/java/oop/course/client/Response.java
@@ -1,5 +1,0 @@
-package oop.course.client;
-
-public interface Response {
-
-}
