+++ src/main/java/oop/course/client/actions/EmptyAction.java
@@ -1,4 +1,7 @@
-package oop.course.client;
+package oop.course.client.actions;
+
+import oop.course.client.actions.Action;
+import oop.course.client.views.IView;
 
 import java.util.function.Consumer;
