+++ src/main/java/oop/course/client/gui/TerminalButton.java
@@ -1,9 +1,10 @@
-package oop.course.client;
+package oop.course.client.gui;
 
 import com.googlecode.lanterna.gui2.Button;
 import com.googlecode.lanterna.gui2.Panel;
+import oop.course.client.gui.TerminalGUIElement;
 
-public class TerminalButton implements TerminalGUIElement{
+public class TerminalButton implements TerminalGUIElement {
     private final Button button;
     public TerminalButton(String text, Runnable action){
         button = new Button(text, action);
