+++ src/main/java/oop/course/client/gui/TerminalInputBox.java
@@ -1,4 +1,4 @@
-package oop.course.client;
+package oop.course.client.gui;
 
 public interface TerminalInputBox extends TerminalGUIElement {
     String text();
