+++ src/main/java/oop/course/client/gui/TerminalPasswordBox.java
@@ -1,4 +1,4 @@
-package oop.course.client;
+package oop.course.client.gui;
 
 import com.googlecode.lanterna.gui2.Panel;
 import com.googlecode.lanterna.gui2.TextBox;
