+++ src/main/java/oop/course/client/requests/AuthorizedRequest.java
@@ -1,8 +1,8 @@
-package oop.course.client;
+package oop.course.client.requests;
 
 import java.io.PrintWriter;
 
-public class AuthorizedRequest implements Request{
+public class AuthorizedRequest implements Request {
     private final Request request;
     private final String token;
