+++ src/main/java/oop/course/client/requests/BasicHttpRequest.java
@@ -1,4 +1,4 @@
-package oop.course.client;
+package oop.course.client.requests;
 
 import java.io.PrintWriter;
