+++ src/main/java/oop/course/client/requests/LoginRequest.java
@@ -1,0 +1,22 @@
+package oop.course.client.requests;
+
+import oop.course.client.responses.LoginResponse;
+import oop.course.client.gui.TerminalForm;
+
+import java.io.PrintWriter;
+
+public class LoginRequest implements Request<LoginResponse> {
+    private final Request base;
+
+    public LoginRequest(TerminalForm terminalForm) {
+        base = new JsonRequest(
+                new BasicHttpRequest(Method.POST, "/login"),
+                terminalForm.json()
+        );
+    }
+
+    @Override
+    public void send(PrintWriter printWriter) {
+        base.send(printWriter);
+    }
+}
