+++ src/main/java/oop/course/client/requests/RegisterRequest.java
@@ -1,13 +1,15 @@
-package oop.course.client;
+package oop.course.client.requests;
+
+import oop.course.client.gui.TerminalForm;
 
 import java.io.PrintWriter;
 
-public class LoginRequest implements Request{
+public class RegisterRequest implements Request {
     private final Request base;
 
-    public LoginRequest(TerminalForm terminalForm) {
+    public RegisterRequest(TerminalForm terminalForm) {
         base = new JsonRequest(
-                new BasicHttpRequest(Method.POST, "/login"),
+                new BasicHttpRequest(Method.POST, "/register"),
                 terminalForm.json()
         );
     }
