+++ src/main/java/oop/course/client/responses/LoginResponse.java
@@ -1,0 +1,4 @@
+package oop.course.client.responses;
+
+public class LoginResponse implements Response {
+}
