+++ src/main/java/oop/course/client/responses/Response.java
@@ -1,0 +1,5 @@
+package oop.course.client.responses;
+
+public interface Response {
+
+}
