+++ src/main/java/oop/course/client/views/AccountView.java
@@ -1,9 +1,14 @@
-package oop.course.client;
+package oop.course.client.views;
 
 import com.googlecode.lanterna.gui2.Direction;
 import com.googlecode.lanterna.gui2.LinearLayout;
 import com.googlecode.lanterna.gui2.Panel;
 import com.googlecode.lanterna.gui2.WindowBasedTextGUI;
+import oop.course.client.actions.Action;
+import oop.course.client.actions.ChangeSceneAction;
+import oop.course.client.gui.TerminalButton;
+import oop.course.client.gui.TerminalText;
+import oop.course.client.gui.TerminalWindow;
 
 import java.util.function.Consumer;
