+++ src/main/java/oop/course/client/views/ActionSelectView.java
@@ -1,13 +1,18 @@
-package oop.course.client;
+package oop.course.client.views;
 
 import com.googlecode.lanterna.gui2.Direction;
 import com.googlecode.lanterna.gui2.LinearLayout;
 import com.googlecode.lanterna.gui2.Panel;
 import com.googlecode.lanterna.gui2.WindowBasedTextGUI;
+import oop.course.client.actions.Action;
+import oop.course.client.actions.ChangeSceneAction;
+import oop.course.client.gui.TerminalButton;
+import oop.course.client.gui.TerminalText;
+import oop.course.client.gui.TerminalWindow;
 
 import java.util.function.Consumer;
 
-public class ActionSelectView implements IView{
+public class ActionSelectView implements IView {
     private final Consumer<Action> actionConsumer;
