+++ src/main/java/oop/course/client/views/LoginView.java
@@ -1,4 +1,4 @@
-package oop.course.client;
+package oop.course.client.views;
 
 import com.googlecode.lanterna.gui2.Direction;
 import com.googlecode.lanterna.gui2.LinearLayout;
@@ -6,6 +6,11 @@ import com.googlecode.lanterna.gui2.Panel;
 import com.googlecode.lanterna.gui2.WindowBasedTextGUI;
 import com.googlecode.lanterna.gui2.dialogs.MessageDialog;
 import com.googlecode.lanterna.gui2.dialogs.MessageDialogButton;
+import oop.course.client.actions.Action;
+import oop.course.client.actions.ChangeSceneAction;
+import oop.course.client.gui.*;
+import oop.course.client.requests.LoginRequest;
+import oop.course.client.requests.Request;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
